# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DynamicCommands API",
    "version": "v1"
  },
  "paths": {
    "/api/arguments/{id}": {
      "get": {
        "tags": [
          "Argument"
        ],
        "operationId": "run.GetArgumentWithID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Argument"
                }
              }
            }
          }
        }
      }
    },
    "/api/arguments": {
      "options": {
        "tags": [
          "Argument"
        ],
        "operationId": "run.PrefligtArgument",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Argument"
        ],
        "operationId": "run.GetAllArguments",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Argument"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Argument"
        ],
        "operationId": "run.InsertArgument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Argument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Argument"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Argument"
        ],
        "operationId": "run.UpdateArgument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Argument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Argument"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Argument"
        ],
        "operationId": "run.DeleteArgument",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/argumentTypes": {
      "options": {
        "tags": [
          "ArgumentType"
        ],
        "operationId": "run.PrefligtArgumentTypes",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "ArgumentType"
        ],
        "operationId": "run.GetAllArgumentTypes",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/favorites": {
      "options": {
        "tags": [
          "Favorite"
        ],
        "operationId": "run.PrefligtFavorites",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Favorite"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Favorite"
        ],
        "operationId": "run.GetAllFavoritesWithUserID",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Favorite"
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Favorite"
        ],
        "operationId": "run.UpdateFavorite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Favorite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Favorite"
        ],
        "operationId": "run.DeleteFavorite",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/favorites/userId": {
      "options": {
        "tags": [
          "Favorite"
        ],
        "operationId": "run.PrefligtFavoritesUserId",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Favorite"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Favorite"
        ],
        "operationId": "run.InsertFavorite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Favorite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/stations": {
      "options": {
        "tags": [
          "Station"
        ],
        "operationId": "run.PrefligtStations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Station"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Station"
        ],
        "operationId": "run.GetAllStations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Station"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Station"
        ],
        "operationId": "run.InsertStation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StationForCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Station"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Station"
        ],
        "operationId": "run.UpdateStation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Station"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Station"
        ],
        "operationId": "run.DeleteStation",
        "parameters": [
          {
            "name": "stationId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/settings/KeyData": {
      "get": {
        "tags": [
          "Setting"
        ],
        "operationId": "run.GetSettingWithKey",
        "parameters": [
          {
            "name": "KeyData",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Setting"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings": {
      "options": {
        "tags": [
          "Setting"
        ],
        "operationId": "run.PrefligtSettings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Setting"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Setting"
        ],
        "operationId": "run.GetAllSettings",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Station"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Setting"
        ],
        "operationId": "run.InsertSetting",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Setting"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Setting"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Setting"
        ],
        "operationId": "run.UpdateSetting",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Setting"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Setting"
        ],
        "operationId": "run.DeleteSetting",
        "parameters": [
          {
            "name": "KeyData",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/commands/moduleId": {
      "get": {
        "tags": [
          "Command"
        ],
        "operationId": "run.GetCommandWithModuleID",
        "parameters": [
          {
            "name": "moduleId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Command"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/modules/subSystemId": {
      "get": {
        "tags": [
          "Module"
        ],
        "operationId": "run.GetModuleWithSubSystemID",
        "parameters": [
          {
            "name": "subSystemId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Module"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/parameters/commandId": {
      "get": {
        "tags": [
          "Parameter"
        ],
        "operationId": "run.GetParameterWithCommandID",
        "parameters": [
          {
            "name": "commandId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Parameter"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/parametertypes": {
      "get": {
        "tags": [
          "ParameterType"
        ],
        "operationId": "run.GetAllParameterTypes",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/satellites": {
      "get": {
        "tags": [
          "Satellite"
        ],
        "operationId": "run.GetAllSatellites",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Satellite"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/subsystems/satelliteId": {
      "get": {
        "tags": [
          "Subsystem"
        ],
        "operationId": "run.GetSubsystemWithSatelliteID",
        "parameters": [
          {
            "name": "satelliteId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Subsystem"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Argument": {
        "type": "object",
        "properties": {
          "idArgument": {
            "type": "integer",
            "format": "int32"
          },
          "labelArgument": {
            "type": "string",
            "nullable": true
          },
          "argumentTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "valueTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "tbValues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Values"
            },
            "nullable": true
          }
        }
      },
      "Favorite": {
        "type": "object",
        "properties": {
          "idFavorites": {
            "type": "integer",
            "format": "int32"
          },
          "commandName": {
            "type": "string",
            "nullable": true
          },
          "dateTimeFavorite": {
            "type": "string",
            "format": "date-time"
          },
          "commandId": {
            "type": "integer",
            "format": "int32"
          },
          "obcid": {
            "type": "integer",
            "format": "int32"
          },
          "parameters": {
            "type": "string",
            "nullable": true
          },
          "isOnline": {
            "type": "boolean"
          },
          "executaionDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ownerUserId": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Command": {
        "type": "object",
        "properties": {
          "idCommand": {
            "type": "integer",
            "format": "int32"
          },
          "commandName": {
            "type": "string",
            "nullable": true
          },
          "obcid": {
            "type": "integer",
            "format": "int32"
          },
          "lenght": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isEnabled": {
            "type": "boolean"
          },
          "moduleId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Module": {
        "type": "object",
        "properties": {
          "idModule": {
            "type": "integer",
            "format": "int32"
          },
          "moduleName": {
            "type": "string",
            "nullable": true
          },
          "obcid": {
            "type": "string",
            "nullable": true
          },
          "subsystemId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "Parameter": {
        "type": "object",
        "properties": {
          "idParameters": {
            "type": "integer",
            "format": "int32"
          },
          "defaultValue": {
            "type": "string",
            "nullable": true
          },
          "formula": {
            "type": "string",
            "nullable": true
          },
          "minValue": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "maxValue": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "commandId": {
            "type": "integer",
            "format": "int32"
          },
          "argumentId": {
            "type": "integer",
            "format": "int32"
          },
          "labelArgument": {
            "type": "string",
            "nullable": true
          },
          "argumentTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "valueTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Values"
            },
            "nullable": true
          }
        }
      },
      "Satellite": {
        "type": "object",
        "properties": {
          "idSatellite": {
            "type": "integer",
            "format": "int32"
          },
          "satelliteName": {
            "type": "string",
            "nullable": true
          },
          "obcid": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Subsystem": {
        "type": "object",
        "properties": {
          "idSubsystem": {
            "type": "integer",
            "format": "int32"
          },
          "subsystemName": {
            "type": "string",
            "nullable": true
          },
          "obcid": {
            "type": "string",
            "nullable": true
          },
          "satelliteId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "ValueType": {
        "type": "object",
        "properties": {
          "idValueType": {
            "type": "integer",
            "format": "int32"
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Values": {
        "type": "object",
        "properties": {
          "idValue": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "parameterVisible": {
            "type": "string",
            "nullable": true
          },
          "parametersId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "Station": {
        "type": "object",
        "properties": {
          "idStation": {
            "type": "integer",
            "format": "int32"
          },
          "stationName": {
            "type": "string",
            "nullable": true
          },
          "stationIp": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "StationForCreate": {
        "type": "object",
        "properties": {
          "stationName": {
            "type": "string",
            "nullable": true
          },
          "stationIp": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Setting": {
        "type": "object",
        "properties": {
          "idSetting": {
            "type": "integer",
            "format": "int32"
          },
          "keyData": {
            "type": "string",
            "nullable": true
          },
          "valueData": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  }
}